fix(web): extract ssr:false dynamic imports into client wrappers
Next.js 16 disallows next/dynamic with ssr:false in Server Components.
Extract SceneCanvas + CustomCursor into ClientEnhancements, and
Preloader + CinematicIntro into ClientShell. Keeps page.tsx and
layout.tsx as Server Components for SSR-first rendering.
