+++ apps/web/src/app/(marketing)/layout.tsx
@@ -1,21 +1,9 @@
-import dynamic from 'next/dynamic';
 import { LenisProvider } from '@/providers/LenisProvider';
 import { ThreeProvider } from '@/providers/ThreeProvider';
 import { Header } from '@/components/layout/Header';
 import { Footer } from '@/components/layout/Footer';
 import { ScrollProgressIndicator } from '@/components/ui/ScrollProgressIndicator';
-
-const SceneCanvas = dynamic(() => import('@/three/canvas/SceneCanvas'), {
-  ssr: false,
-});
-
-const CustomCursor = dynamic(
-  () =>
-    import('@/components/ui/CustomCursor/CustomCursor').then(
-      (m) => m.CustomCursor,
-    ),
-  { ssr: false },
-);
+import { ClientEnhancements } from '@/components/layout/ClientEnhancements';
 
 export default function MarketingLayout({
   children,
@@ -27,8 +15,7 @@ export default function MarketingLayout({
       <ThreeProvider>
         <ScrollProgressIndicator />
         <Header />
-        <SceneCanvas />
-        <CustomCursor />
+        <ClientEnhancements />
         <main style={{ paddingTop: 'var(--header-height)' }}>{children}</main>
         <Footer />
       </ThreeProvider>
