+++ apps/web/src/app/(marketing)/page.tsx
@@ -1,28 +1,11 @@
-import dynamic from 'next/dynamic';
+import { ClientShell } from '@/features/homepage/ClientShell';
 import { HeroSection } from '@/features/homepage/HeroSection/HeroSection';
 import { SectionTransition } from '@/components/ui';
-
-const Preloader = dynamic(
-  () =>
-    import('@/features/homepage/Preloader/Preloader').then(
-      (m) => m.Preloader,
-    ),
-  { ssr: false },
-);
-
-const CinematicIntro = dynamic(
-  () =>
-    import('@/features/homepage/CinematicIntro/CinematicIntro').then(
-      (m) => m.CinematicIntro,
-    ),
-  { ssr: false },
-);
 
 export default function HomePage() {
   return (
     <>
-      <Preloader />
-      <CinematicIntro />
+      <ClientShell />
       <HeroSection />
       <SectionTransition from="blue" to="white" />
     </>
